Be able to compile a static libc for myos
This required some erroneous changes to eliminate build errors, for
example:

  * removing ABI checks of fast types (types.h),

  * removing other static size asserts (options/rtdl/generic/*.cpp),

  * defining SSIZE_MAX, NAME_MAX and PATH_MAX by hand in options/posix/,
    because for some reason they were not defined by including limits.h.

My guess is that the build errors are caused by the fact that mlibc does
not have full support of x86.

Of course, being able to compile mlibc does not mean it will work at all
with myos.  However, it's a place to start.